Started to write a small editor to enable the user to add points from a visual interface

+++ src/algo/corpus.cpp
@@ -188,9 +188,9 @@ void Corpus::display()
     }
 }
 
-vector<int> Corpus::bounds()
+vector<float> Corpus::bounds()
 {
-    vector<int> result(mDimension*2,0);
+    vector<float> result(mDimension*2,0);
 
     for(int i = 0; i != mSize; ++i)
         for(int j = 0; j != mDimension; ++j)
